requirements.txt
# Core dependencies for all services
boto3>=1.34.0
fastapi>=0.109.0
uvicorn>=0.27.0
watchdog>=3.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
httpx>=0.26.0
python-magic>=0.4.27
mcp>=0.9.0
jinja2>=3.1.3
PyYAML>=6.0.1
pandas>=2.0.0

# OCR dependencies (Tesseract)
pytesseract>=0.3.10
Pillow>=10.0.0

# PostgreSQL database drivers (replacing DuckDB)
asyncpg>=0.29.0          # Async PostgreSQL driver (primary)
psycopg2-binary>=2.9.9   # Sync PostgreSQL driver (for migrations/scripts)

# Optional: SQLAlchemy for ORM support
# sqlalchemy>=2.0.23
# alembic>=1.13.0        # Database migrations

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
requests>=2.32.0            # For E2E integration tests
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0